refactor: support prefix and slash features
